Implement fast stop of the kernel build process
This commit fixes the issue #2.

Stopping of the docker container using SIGTERM is not fast and stable.
So I decided to use `docker run` with `--cidfile` for storing
the container id.

Now I catch the KeyboardInterrupt in python and execute
finish_container.sh, which kills the container and removes the cidfile.

In case of normal build without KeyboardInterrupt, finish_container.sh
checks that the container is happily finished on its own and removes
the cidfile as well.