Give user a choice to choose which container to build
Created a fucntion for every container build, user passes the argument
and builds only the version user asked for.
This saves a lot of time and space on the disk

If no arguments are provided all the containers will be built. I kept
it that way since I didn't wanna change the legacy behaviour.

Signed-off-by: Mukesh Kumar Chaurasiya <mchauras@linux.ibm.com>