Dockerfile: Add rsync to the container
When working with linux, kernel headers are often required when
building out of tree modules or various userspace components, or the
kernel selftests.

The kernel has a makefile target `headers' for this [1].
However, this makefile target uses `rsync`, which is not currently part
of the container.

Add rsync to the installed binaries of the containers, to support
installing linux headers.

[1] https://www.kernel.org/doc/html/latest/dev-tools/kselftest.html

Signed-off-by: Evangelos Petrongonas <vpetrog@ieee.org>