build_linux: Remove manual shell redirection
If the container is not running in interactive mode, '2>&1' is appended
to the end of the container command. This is problematic. When
`subprocess.Popen()` is called with a list of arguments, each item on
the list is treated as a literal argument, not as shell syntax.
Therefore, when passing the following arguments for example:
```
-- -C /src/tools/testing/selftests TARGETS=kvm
```
it translates to
```
make O=../out/ -j 12 -C /src/tools/testing/selftests TARGETS=kvm 2>&1"
```
which fails as make can not find a target for `2>&1`.

Remove this cmdline extension as a) it is causing build issues b) is
unnecessary as the redirection is already handled with the `Popen`
`stderr=subprocess.STDOUT` parameter.

Signed-off-by: Evangelos Petrongonas <vpetrog@ieee.org>